# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SEOaxe API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deploy/wordpress:
    post:
      operationId: deployToWordpress
      tags: [seo]
      summary: Push optimized HTML to a WordPress page or post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WordpressDeployRequest"
      responses:
        "200":
          description: Deployed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Deploy failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /deploy/shopify:
    post:
      operationId: deployToShopify
      tags: [seo]
      summary: Push optimized HTML to a Shopify page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopifyDeployRequest"
      responses:
        "200":
          description: Deployed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeployResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Deploy failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /sitemap-urls:
    get:
      operationId: listSitemapUrls
      tags: [seo]
      summary: List sitemap entries
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SitemapUrlListResult"
    post:
      operationId: addSitemapUrl
      tags: [seo]
      summary: Add a URL to the sitemap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SitemapUrlInput"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SitemapUrlRecord"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sitemap-urls/{id}:
    delete:
      operationId: deleteSitemapUrl
      tags: [seo]
      summary: Remove a URL from the sitemap
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteOptimizationResponseAlias"
...
  /monitor/sites/{id}:
    delete:
      operationId: deleteMonitoredSite
      tags: [seo]
      summary: Remove a monitored site
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    WordpressDeployRequest:
      type: object
      properties:
        siteUrl:
          type: string
          format: uri
        username:
          type: string
        appPassword:
          type: string
        postType:
          type: string
          enum: [pages, posts]
        postId:
          type: integer
        html:
          type: string
      required:
        - siteUrl
        - username
        - appPassword
        - postType
        - postId
        - html
    ShopifyDeployRequest:
      type: object
      properties:
        shop:
          type: string
        accessToken:
          type: string
        pageId:
          type: integer
        html:
          type: string
      required:
        - shop
        - accessToken
        - pageId
        - html
    DeployResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        url:
          type: string
          nullable: true
      required:
        - success
        - message
...
    SitemapUrlRecord:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        priority:
          type: integer
        changefreq:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - url
        - priority
        - changefreq
        - createdAt
    SitemapUrlInput:
      type: object
      properties:
        url:
          type: string
          format: uri
        priority:
          type: integer
          minimum: 0
          maximum: 100
        changefreq:
          type: string
          enum: [always, hourly, daily, weekly, monthly, yearly, never]
      required:
        - url
    SitemapUrlListResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SitemapUrlRecord"
      required:
        - items
    DeleteOptimizationResponseAlias:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    SimpleOk:
      type: object
      properties:
        ok: { type: boolean }
        message: { type: string }
      required: [ok]
...