# lib/api-zod/src/generated/types/sitemapUrlListResult.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * SEOaxe API
 * OpenAPI spec version: 0.1.0
 */
import type { SitemapUrlRecord } from './sitemapUrlRecord';

export interface SitemapUrlListResult {
  items: SitemapUrlRecord[];
}