# lib/api-zod/src/generated/types/sitemapUrlRecord.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * SEOaxe API
 * OpenAPI spec version: 0.1.0
 */

export interface SitemapUrlRecord {
  id: number;
  url: string;
  priority: number;
  changefreq: string;
  createdAt: Date;
}